コマンドライン戻り値の不具合を修正した
