--- AILZ80ASM.Test/LBAssembleTest.cs
@@ -1,51 +1,0 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Reflection;
-
-namespace AILZ80ASM.Test
-{
-    [TestClass]
-    public class LBAssembleTest
-    {
-        [TestMethod]
-        public void TestLB_CALL_Address()
-        {
-            Lib.Assemble_AreSame(MethodBase.GetCurrentMethod().Name);
-        }
-
-        [TestMethod]
-        public void TestLB_DBDW_Test()
-        {
-            Lib.Assemble_AreSame(MethodBase.GetCurrentMethod().Name);
-        }
-
-        [TestMethod]
-        public void TestLB_DSDBSDWS_Test()
-        {
-            Lib.Assemble_AreSame(MethodBase.GetCurrentMethod().Name);
-        }
-
-        [TestMethod]
-        public void TestLB_DSDBSDWS_Trim_Test()
-        {
-            Lib.Assemble_AreSame(MethodBase.GetCurrentMethod().Name, true);
-        }
-
-        [TestMethod]
-        public void TestLB_EQU_Test()
-        {
-            Lib.Assemble_AreSame(MethodBase.GetCurrentMethod().Name);
-        }
-
-        [TestMethod]
-        public void TestLB_JP_And_JR()
-        {
-            Lib.Assemble_AreSame(MethodBase.GetCurrentMethod().Name);
-        }
-
-        [TestMethod]
-        public void TestLB_LD_Address()
-        {
-            Lib.Assemble_AreSame(MethodBase.GetCurrentMethod().Name);
-        }
-    }
-}
