--- AILZ80ASM.Test/LNAssembleTest.cs
@@ -1,15 +1,0 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Reflection;
-
-namespace AILZ80ASM.Test
-{
-    [TestClass]
-    public class LNAssembleTest
-    {
-        [TestMethod]
-        public void TestLN_Comment()
-        {
-            Lib.Assemble_AreSame(MethodBase.GetCurrentMethod().Name);
-        }
-    }
-}
