+++ AILZ80ASM.Test/Test/TestSS_Main/Error.Z80
@@ -1,0 +1,1 @@
+﻿	ld a, g
