+++ AILZ80ASM.Test/Test/TestSS_Main/Success.Z80
@@ -1,0 +1,1 @@
+﻿	ld a, 0
