+++ AILZ80ASM.Test/Z80RDAssembleTest.cs
@@ -1,0 +1,15 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.Reflection;
+
+namespace AILZ80ASM.Test
+{
+    [TestClass]
+    public class Z80RDAssembleTest
+    {
+        [TestMethod]
+        public void TestRD_ALL()
+        {
+            Lib.Assemble_AreSame(MethodBase.GetCurrentMethod().Name);
+        }
+    }
+}
