+++ AILZ80ASM/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "AILZ80ASM": {
       "commandName": "Project",
-      "commandLineArgs": ".\\SoundTestV2.z80 -cmt -l -lm simple -t -e"
+      "commandLineArgs": ".\\angle.z80"
     }
   }
 }
