+++ README.md
@@ -59,6 +59,12 @@ AILZ80ASM [<オプション>] <オプション指定文字列:ファイル名等
 > AILZ80ASM -h -om
 ```
 
+## コマンドライン 戻り値
+- 0: アセンブルが正常に終了
+- 1: アセンブルにエラーがあり、処理途中で終了
+- 2: コマンドライン引数に間違いがあり、終了
+- 3: 内部エラーが発生、アセンブルの強制終了
+
 # ソースコード書式
 
 ## ニーモニック
